feat: 더미 데이터 삽입을 위해 repo 구현

+++ src/main/java/com/example/be/concert/repository/HallTemplateRepository.java
@@ -3,4 +3,4 @@ package com.example.be.concert.repository;
 import com.example.be.concert.domain.HallTemplate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HallTemplateRepository extends JpaRepository<HallTemplate, Integer> {}
+public interface HallTemplateRepository extends JpaRepository<HallTemplate, Long> {}
